UPDATE: PrimesTableConsole #1 Main controller
create PrimesTableConsole,
move validation logic from PrimesTable service to the main controller
