+++ Classlib/Test/Services/PrimesTableTest.php
@@ -2,8 +2,6 @@
 namespace DevSpace\Test\Services;
 
 use DevSpace\Services\PrimesTable;
-use DevSpace\Interfaces\Resources\IMessages;
-use DevSpace\Interfaces\Validators\INaturalNumber;
 use DevSpace\Interfaces\MathLib\IPrimes;
 use DevSpace\Interfaces\MathLib\ITimesTable;
 use DevSpace\Test\Core\TestCase;
@@ -14,12 +12,6 @@ class PrimesTableTest extends TestCase
     /** @var  PrimesTable */
     private $subject;
 
-    /** @var  IMessages | PHPUnit_Framework_MockObject_MockObject */
-    private $mockMessages;
-
-    /** @var  INaturalNumber | PHPUnit_Framework_MockObject_MockObject */
-    private $mockSizeValidator;
-
     /** @var  IPrimes | PHPUnit_Framework_MockObject_MockObject */
     private $mockPrimesGenerator;
 
@@ -27,13 +19,9 @@ class PrimesTableTest extends TestCase
     private $mockTimesTableGenerator;
 
     public function setUp() {
-        $this->mockMessages = $this->getMock(IMessages::class);
-        $this->mockSizeValidator = $this->getMock(INaturalNumber::class);
         $this->mockPrimesGenerator = $this->getMock(IPrimes::class);
         $this->mockTimesTableGenerator = $this->getMock(ITimesTable::class);
         $this->subject = new PrimesTable(
-            $this->mockMessages,
-            $this->mockSizeValidator,
             $this->mockPrimesGenerator,
             $this->mockTimesTableGenerator
         );
@@ -43,21 +31,12 @@ class PrimesTableTest extends TestCase
     {
         $this->assertInstanceOf('DevSpace\Interfaces\Services\IPrimesTable', $this->subject);
     }
-    
-    public function testGetPrimesTableDisplaysOutputMessagesIfSizeArgInValid()
-    {
-        $expectedResult = 'Help';
-        $this->expectSizeArgValidated(false);
-        $this->expectToGetAMessage('MSG_PRMGEN_INVALID_INPUT', $expectedResult);
-        $this->assertEquals($expectedResult, $this->subject->getPrimesTable(-1));
-    }
 
     public function testGetPrimesTableWillGeneratePrimes()
     {
         $primes = array('i am a fake array of n prime numbers');
         $timesTable = array('i am a times table of given primes');
         $size = 4;
-        $this->expectSizeArgValidated(true);
         $this->expectGeneratingPrimes($size, $primes);
         $this->expectToGetTimesTableOfPrimes($primes, $timesTable);
         $this->assertEquals($timesTable, $this->subject->getPrimesTable($size));
@@ -79,23 +58,5 @@ class PrimesTableTest extends TestCase
             ->method('primes')
             ->with($arg)
             ->willReturn($result);
-    }
-
-    private function expectSizeArgValidated($expected)
-    {
-        $this->mockSizeValidator
-            ->expects($this->once())
-            ->method('validate')
-            ->withAnyParameters()
-            ->willReturn($expected);
-    }
-
-    public function expectToGetAMessage($expectedArg, $expectedMessage)
-    {
-        $this->mockMessages
-            ->expects($this->once())
-            ->method('getMessage')
-            ->with($expectedArg)
-            ->willReturn($expectedMessage);
     }
 }
